+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - turrent lamps cast with the media of the item
 - added Media Reflection - gets media of casting environment
 - added Media Purification - gets media of player, allay, impetus, or item on floor
+- added Construct spell - places a specific block type from your inventory with parameters for orientation and horizontal vector
 
 ## 1.5.0
 - added spell to conjure llama spit